requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.7
bcrypt==3.1.7
beautifulsoup4==4.9.1
bootstrap4==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-mediumeditor==1.0.0
django-registration==3.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
Faker==4.1.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
wincertstore==0.2
